discordPoints/requirements.txt
gspread==0.6.2
oauth2client==4.1.2
discord.py==0.16.7